Created IO library for Leo
